+++ src/post.js
@@ -10,12 +10,56 @@
                              missing a title.
  */
 Module['parse'] = function(blueprint, options, callback) {
+  if (arguments.length < 1 || arguments.length > 3) {
+    throw new TypeError('wrong number of arguments, `parse(string, options, callback)` expected');
+  }
+
+  if (typeof blueprint !== 'string') {
+    throw new TypeError('wrong 1st argument - string expected, `parse(string, options, [callback])`')
+  }
+
+  if (options && typeof options === 'function') {
+    callback = options;
+    options = {};
+  }
+
+  options = options || {};
+
+  if (typeof options !== 'object') {
+    throw new TypeError('wrong 2nd argument - object expected, `parse(string, options, [callback])`');
+  }
+
+  if (callback && typeof callback !== 'function') {
+    throw new TypeError('wrong number of arguments, `parse(string, options)` expected');
+  }
+
+  var sync = false;
+
+  if (options.sync) {
+    sync = true;
+    delete options.sync;
+  }
+
+  var allowedOptions = ['generateSourceMap', 'exportSourcemap', 'requireBlueprintName'];
+
+  Object.keys(options).forEach(function (key) {
+    if (allowedOptions.indexOf(key) === -1) {
+      throw new TypeError('unrecognized option \'' + key + '\', expected: \'requireBlueprintName\', \'generateSourceMap\'');
+    }
+  });
+
   if (false === this.ready) {
     var err = new Error('Module not ready!');
-    if (callback) {
-      return callback(err, null);
+
+    if (sync) {
+      throw err;
     }
-    return err;
+
+    if (callback) {
+      return callback(err);
+    }
+
+    return Promise.reject(err);
   }
 
   try {
@@ -28,14 +72,12 @@ Module['parse'] = function(blueprint, options, callback) {
 
     stringToUTF8(blueprint, buffer, bufferLen);
 
-    if (options) {
-      if (options.generateSourceMap) {
-        sourcemap = options.generateSourceMap;
-      }
+    if (options.generateSourceMap || options.exportSourcemap) {
+      sourcemap = true;
+    }
 
-      if (options.requireBlueprintName) {
-        requireBlueprintName = options.requireBlueprintName;
-      }
+    if (options.requireBlueprintName) {
+      requireBlueprintName = true;
     }
 
     var res = _c_parse(buffer, requireBlueprintName, sourcemap, chptr);
@@ -49,27 +91,62 @@ Module['parse'] = function(blueprint, options, callback) {
     _free(chptr);
 
   } catch (ex) {
+    if (sync) {
+      throw ex;
+    }
 
     if (callback) {
-      return callback(ex, null);
+      return callback(ex);
     }
-    throw ex;
+
+    return Promise.reject(ex);
   }
 
-  result = (options && options.json === false) ? output : JSON.parse(output);
+  var error = null;
+  var result = null;
+
+  if (res < 0) {
+    error = new Error('Parser: Unknown Error');
+  } else {
+    result = (options.json === false) ? output : JSON.parse(output);
+  }
+
+  if (sync) {
+    if (error) {
+      throw error;
+    }
+
+    return result;
+  }
 
   if (callback) {
-    if (res) {
-      return callback(result, result);
-    } else {
-      return callback(null, result);
-    }
+    return callback(error, result);
   }
 
-  return result;
+  if (error) {
+    return Promise.reject(error);
+  }
+
+  return Promise.resolve(result);
 };
 
 Module['parseSync'] = function(blueprint, options) {
+  if (arguments.length < 1 || arguments.length > 2) {
+    throw new TypeError('wrong number of arguments, `parseSync(string, options)` expected');
+  }
+
+  if (typeof blueprint !== 'string') {
+    throw new TypeError('wrong 1st argument - string expected, `parseSync(string, options)`')
+  }
+
+  options = options || {};
+
+  if (typeof options !== 'object') {
+    throw new TypeError('wrong 2nd argument - object expected, `parseSync(string, options)`');
+  }
+
+  options.sync = true;
+
   return Module.parse(blueprint, options);
 };
 
@@ -83,12 +160,56 @@ Module['parseSync'] = function(blueprint, options) {
                              missing a title.
  */
 Module['validate'] = function(blueprint, options, callback) {
+  if (arguments.length < 1 || arguments.length > 3) {
+    throw new TypeError('wrong number of arguments, `validate(string, options, callback)` expected');
+  }
+
+  if (typeof blueprint !== 'string') {
+    throw new TypeError('wrong 1st argument - string expected, `validate(string, options, [callback])`')
+  }
+
+  if (options && typeof options === 'function') {
+    callback = options;
+    options = {};
+  }
+
+  options = options || {};
+
+  if (typeof options !== 'object') {
+    throw new TypeError('wrong 2nd argument - object expected, `validate(string, options, [callback])`');
+  }
+
+  if (callback && typeof callback !== 'function') {
+    throw new TypeError('wrong number of arguments, `validate(string, options)` expected');
+  }
+
+  var sync = false;
+
+  if (options.sync) {
+    sync = true;
+    delete options.sync;
+  }
+
+  var allowedOptions = ['requireBlueprintName'];
+
+  Object.keys(options).forEach(function (key) {
+    if (allowedOptions.indexOf(key) === -1) {
+      throw new TypeError('unrecognized option \'' + key + '\', expected: \'requireBlueprintName\'');
+    }
+  });
+
   if (false === this.ready) {
     var err = new Error('Module not ready!');
-    if (callback) {
-      return callback(err, null);
+
+    if (sync) {
+      throw err;
     }
-    return err;
+
+    if (callback) {
+      return callback(err);
+    }
+
+    return Promise.reject(err);
   }
 
   try {
@@ -99,8 +220,8 @@ Module['validate'] = function(blueprint, options, callback) {
     var requireBlueprintName = false;
     var output = null;
 
-    if (options && options.requireBlueprintName) {
-      requireBlueprintName = options.requireBlueprintName;
+    if (options.requireBlueprintName) {
+      requireBlueprintName = true;
     }
 
     stringToUTF8(blueprint, buffer, bufferLen);
@@ -110,25 +231,67 @@ Module['validate'] = function(blueprint, options, callback) {
 
     if (res) {
       var ptrstr = getValue(chptr, '*');
-      output = (options && options.json === false) ? Pointer_stringify(ptrstr) : JSON.parse(Pointer_stringify(ptrstr));
+      output = (options.json === false) ? Pointer_stringify(ptrstr) : JSON.parse(Pointer_stringify(ptrstr));
       _free(ptrstr);
     }
     _free(chptr);
 
   } catch (ex) {
+    if (sync) {
+      throw ex;
+    }
 
     if (callback) {
-      return callback(ex, null);
+      return callback(ex);
     }
-    throw ex;
+
+    return Promise.reject(ex);
+  }
+
+  var error = null;
+  var result = null;
+
+  if (res < 0) {
+    error = new Error('Parser: Unknown Error');
+  } else {
+    result = output;
+  }
+
+  if (sync) {
+    if (error) {
+      throw error;
+    }
+
+    return result;
   }
 
   if (callback) {
-    return callback(null, output);
+    return callback(error, result);
   }
-  return output;
+
+  if (error) {
+    return Promise.reject(error);
+  }
+
+  return Promise.resolve(result);
 };
 
 Module['validateSync'] = function(blueprint, options) {
+  if (arguments.length < 1 || arguments.length > 2) {
+    throw new TypeError('wrong number of arguments, `validateSync(string, options)` expected');
+  }
+
+  if (typeof blueprint !== 'string') {
+    throw new TypeError('wrong 1st argument - string expected, `validateSync(string, options)`')
+  }
+
+  options = options || {};
+
+  if (typeof options !== 'object') {
+    throw new TypeError('wrong 2nd argument - object expected, `validateSync(string, options)`');
+  }
+
+  options.sync = true;
+
   return Module.validate(blueprint, options);
 };