+++ test/await-test.js
@@ -1,0 +1,4 @@
+const awaitHelper = require('../ext/protagonist/test/helpers/await');
+var drafter = require('../lib/drafter.nomem.js');
+
+awaitHelper(drafter);