+++ test/callback-test.js
@@ -1,0 +1,4 @@
+const callbackHelper = require('../ext/protagonist/test/helpers/callback');
+var drafter = require('../lib/drafter.nomem.js');
+
+callbackHelper(drafter);