+++ test/params-test.js
@@ -1,0 +1,4 @@
+const paramsHelper = require('../ext/protagonist/test/helpers/params');
+var drafter = require('../lib/drafter.nomem.js');
+
+paramsHelper(drafter);