+++ test/promise-test.js
@@ -1,0 +1,4 @@
+const promiseHelper = require('../ext/protagonist/test/helpers/promise');
+var drafter = require('../lib/drafter.nomem.js');
+
+promiseHelper(drafter);